Fix invalid Content-Length if set in handler
If a handler later in the chain sets the Content-Length header, removing
it afterwards is already too late. The header was already sent the
client.
This change will prevent to write the header in all next handlers, so
that the compressed body doesn't have a different size than
Content-Length.
Browsers like Safari (but also Firefox and Chrome for mobile) will
report connection errors instead of displaying the content of the page.

closes #14